zsh/_meson: complete build options from introspection data
This uses an external Python script which parses the JSON output
of `meson introspect`. It is intended to be installed alongside
the main completion function.

The completion can be customized with the usual zsh tools, documented in
the zshcompsys manual; for example, the following command specifies that
completion for `meson configure` in the option name position after `-D`
offers project-defined build options and compiler options, in that
order, and only offers the rest if no completions are available:

    zstyle ':completion:*:complete:meson-configure:option-D-1:*' tag-order 'user-options base-options'

Since we parse introspection data, we do not need any knowledge of the
various builtin options and whether they are applicable to the toolchain
being used, or if they come and go in future versions of Meson. That
means the completion results likely are as complete as possible, with
the maintenance burden greatly reduced.

To work correctly with `meson setup -D`, when a build directory is not
available, we call `meson introspect meson.build` on the top-level meson
build file. This is 20-30 times slower than introspecting a configured
build directory, but at least it works.

The `_meson` completion function can no longer be copied into the
vendor completion directory as is, but should be put through a
substitution pass by packagers, for example:

    sed "s!@ZSH_SITE_FUNCTIONS@!/usr/share/zsh/site-functions!" data/shell-completions/zsh/_meson.in > /usr/share/zsh/site-functions/_meson
    install -Dpm644 data/shell-completions/zsh/meson-comp-helper.py /usr/share/zsh/site-functions/

The installation paths may differ.

NB: the functions are regenerated on each invocation, and a caching
strategy has to support different cached content for different projects
and builddirs. This is an even bigger latency problem if the zsh user
has configured the completion system to try different matching
strategies and e. g. correcting completion, `meson introspect` and the
Python generator will be called 2 to 5 times in that case.
The solution would be to use the `_store_cache` primitive and configure
it with `zstyle`.